refactor(devtoolkit): remove cohort-gated plan API
DevToolkit has no deployed callers, so retaining plan/apply would create two competing environment models without migration value.

Remove the plan/apply surface and its cohort-bound planner, doctor, target, handoff, receipt, and model layers. Keep cohort files only as optional qualification inputs to capability resolution, and preserve generic wheel-platform coverage in a dedicated test module.

BREAKING CHANGE: remove DevToolkit.plan(), DevToolkit.apply(), PrepareRequest, PreparationPlan, PrepareResult, and the legacy target and handoff APIs.
