--- scripts/devToolkit/trtmc_devtoolkit/handoff.py
@@ -1,8 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-"""Backward-compatible imports for handoff recipes."""
-
-from .recipes.handoffs import performance_handoff, profiling_handoff, validation_handoff
-
-__all__ = ["performance_handoff", "profiling_handoff", "validation_handoff"]
