+++ scripts/devToolkit/trtmc_devtoolkit/models.py
@@ -1,143 +1,15 @@
 # SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-"""Immutable request, plan, result, and environment-contract models."""
+"""Shared capability errors and observed toolchain facts."""
 
 from __future__ import annotations
 
-from dataclasses import asdict, dataclass, field
-from pathlib import Path
-from typing import Literal
-
-
-PrepareMode = Literal["development", "installed"]
-CohortStatus = Literal["supported", "experimental"]
-LocalDependencyMode = Literal["managed", "system"]
-TargetKind = Literal["local", "docker"]
+from dataclasses import dataclass
 
 
 class DevToolkitError(RuntimeError):
     """A user-facing environment preparation error."""
-
-
-@dataclass(frozen=True)
-class PackagePin:
-    name: str
-    version: str
-
-
-@dataclass(frozen=True)
-class DownloadArtifact:
-    url: str
-    sha256: str
-
-
-@dataclass(frozen=True)
-class ManagedLocalContract:
-    python_packages: tuple[PackagePin, ...]
-
-
-@dataclass(frozen=True)
-class ArchitectureContract:
-    dockerfile: str
-    docker_context: str
-    container_python_version: str
-    wheel_platform: str
-    tensorrt_include_dir: str
-    tensorrt_library_dir: str
-    tensorrt_headers: DownloadArtifact
-
-
-@dataclass(frozen=True)
-class EnvironmentCohort:
-    schema_version: int
-    id: str
-    status: CohortStatus
-    targets: tuple[TargetKind, ...]
-    tensorrt_version: str
-    tensorrt_apt_version: str
-    cuda_version: str
-    python_versions: tuple[str, ...]
-    architectures: dict[str, ArchitectureContract]
-    managed_local: ManagedLocalContract
-    source: Path = field(compare=False)
-
-
-@dataclass(frozen=True)
-class DockerTarget:
-    """Prepare a persistent development container backed by a cohort image."""
-
-    gpu: str = "0"
-    image: str | None = None
-    container_name: str | None = None
-    hf_cache: str | None = None
-    storage_root: str | None = None
-    forward_environment: tuple[str, ...] = (
-        "HF_TOKEN",
-        "HUGGING_FACE_HUB_TOKEN",
-        "HF_HUB_DISABLE_XET",
-        "HF_HUB_DOWNLOAD_TIMEOUT",
-        "HF_HUB_ETAG_TIMEOUT",
-    )
-    rebuild_image: bool = False
-    kind: Literal["docker"] = "docker"
-
-
-@dataclass(frozen=True)
-class LocalTarget:
-    """Prepare a local venv without modifying system CUDA, TensorRT, or drivers."""
-
-    python: str = "python3"
-    gpu: str = "0"
-    dependency_mode: LocalDependencyMode = "managed"
-    kind: Literal["local"] = "local"
-
-
-@dataclass(frozen=True)
-class PrepareRequest:
-    tensorrt: str
-    cuda: str
-    target: DockerTarget | LocalTarget
-    mode: PrepareMode = "development"
-    python_version: str = "3.12"
-    architecture: str | None = None
-    allow_experimental: bool = False
-    model: object | None = None
-
-
-@dataclass(frozen=True)
-class PlanStep:
-    id: str
-    description: str
-    mutates: bool
-
-
-@dataclass(frozen=True)
-class PreparationPlan:
-    request: PrepareRequest
-    cohort: EnvironmentCohort
-    repository: Path
-    architecture: str
-    source_revision: str
-    environment_id: str
-    run_id: str
-    state_dir: Path
-    image_fingerprint: str | None
-    steps: tuple[PlanStep, ...]
-
-    def as_dict(self) -> dict[str, object]:
-        payload = asdict(self)
-        payload["repository"] = str(self.repository)
-        payload["state_dir"] = str(self.state_dir)
-        payload["cohort"]["source"] = str(self.cohort.source)  # type: ignore[index]
-        return payload
-
-
-@dataclass(frozen=True)
-class ProbeResult:
-    name: str
-    status: Literal["pass", "fail", "warning"]
-    detail: str
 
 
 @dataclass(frozen=True)
@@ -154,32 +26,3 @@ class ToolchainObservation:
     cuda_root: str | None = None
     image_id: str | None = None
     architecture: str | None = None
-
-
-@dataclass(frozen=True)
-class EnvironmentHandle:
-    kind: Literal["docker", "local"]
-    fingerprint: str
-    trtmc: str
-    python: str
-    activate_command: str
-    image_ref: str | None = None
-    container_name: str | None = None
-    environment: dict[str, str] = field(default_factory=dict)
-    observation: ToolchainObservation | None = None
-
-
-@dataclass(frozen=True)
-class PrepareResult:
-    plan: PreparationPlan
-    environment: EnvironmentHandle
-    receipt: Path
-    wheel: Path | None = None
-    bundle: Path | None = None
-
-
-@dataclass(frozen=True)
-class HandoffPlan:
-    name: str
-    command: tuple[str, ...]
-    environment: dict[str, str]
