--- scripts/devToolkit/trtmc_devtoolkit/recipes/__init__.py
@@ -1,8 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-"""Optional workflow recipes built from the DevToolkit capability API."""
-
-from .handoffs import performance_handoff, profiling_handoff, validation_handoff
-
-__all__ = ["performance_handoff", "profiling_handoff", "validation_handoff"]
