--- scripts/devToolkit/trtmc_devtoolkit/targets.py
@@ -1,707 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-"""Local and Docker target provisioning, build, installation, and smoke steps."""
-
-from __future__ import annotations
-
-import json
-import os
-import re
-import shlex
-from pathlib import Path
-
-from .models import (
-    DevToolkitError,
-    EnvironmentHandle,
-    PreparationPlan,
-    ToolchainObservation,
-)
-from .runner import Runner, command_output
-from .toolchain import (
-    CUDA_RELEASE,
-    ManagedLocalProvisioner,
-    observe_local_toolchain,
-    require_toolchain_observation,
-    system_tensorrt_paths,
-    tensorrt_header_version_text,
-)
-
-
-IMAGE_FINGERPRINT_LABEL = "org.nvidia.trtmc.devtoolkit-input-fingerprint"
-ENVIRONMENT_LABEL = "org.nvidia.trtmc.devtoolkit-environment"
-CONTAINER_NAME = re.compile(r"[a-zA-Z0-9][a-zA-Z0-9_.-]+")
-
-
-def _installed_base_requirements(
-    python: Path,
-    runner: Runner,
-    *,
-    repository: Path,
-    environment: dict[str, str],
-) -> list[str]:
-    """Read the editable distribution's requirements without selecting TensorRT wheels."""
-    script = (
-        "import importlib.metadata, json; "
-        "requirements=importlib.metadata.requires('tensorrt-model-connect') or []; "
-        "print(json.dumps([item for item in requirements "
-        "if not item.lstrip().lower().startswith('tensorrt')]))"
-    )
-    raw = command_output(
-        runner,
-        [python, "-c", script],
-        cwd=repository,
-        env=environment,
-    )
-    try:
-        requirements = json.loads(raw)
-    except json.JSONDecodeError as error:
-        raise DevToolkitError(
-            "Could not read TRTMC Python requirements from editable package metadata"
-        ) from error
-    if not isinstance(requirements, list) or not all(
-        isinstance(requirement, str) and requirement.strip() for requirement in requirements
-    ):
-        raise DevToolkitError("TRTMC editable package metadata contains invalid requirements")
-    return requirements
-
-
-def _python_tag(version: str) -> str:
-    return "py" + version.replace(".", "")
-
-
-def _container_path(plan: PreparationPlan, path: Path) -> Path:
-    resolved = path.resolve()
-    try:
-        relative = resolved.relative_to(plan.state_dir.resolve())
-        return Path("/trtmc-devtoolkit-run") / relative
-    except ValueError:
-        pass
-    relative = resolved.relative_to(plan.repository.resolve())
-    return Path("/workspace/tensorrt-model-connect") / relative
-
-
-def _docker_exec(
-    container: str,
-    command: list[str | Path],
-    *,
-    environment: dict[str, str] | None = None,
-) -> list[str]:
-    arguments = ["docker", "exec"]
-    for name, value in sorted((environment or {}).items()):
-        arguments.extend(["--env", f"{name}={value}"])
-    arguments.append(container)
-    arguments.extend(str(item) for item in command)
-    return arguments
-
-
-def _write_local_activation(
-    destination: Path,
-    venv: Path,
-    environment: dict[str, str],
-) -> None:
-    lines = [
-        "# Generated by TRTMC devToolkit. Source this file in a POSIX shell.",
-        f". {shlex.quote(str(venv / 'bin' / 'activate'))}",
-    ]
-    lines.extend(
-        f"export {name}={shlex.quote(value)}" for name, value in sorted(environment.items())
-    )
-    destination.write_text("\n".join(lines) + "\n", encoding="utf-8")
-
-
-def _development_runtime_environment(
-    environment: dict[str, str],
-    build: Path,
-) -> dict[str, str]:
-    return {
-        **environment,
-        "PATH": f"{build}:{environment.get('PATH', '')}",
-        "TRTMC_MODEL_PLUGIN_DIR": str(build / "models"),
-    }
-
-
-class DockerEnvironment:
-    def __init__(self, repository: Path, runner: Runner):
-        self.repository = repository.resolve()
-        self.runner = runner
-
-    def prepare(
-        self,
-        plan: PreparationPlan,
-        *,
-        sm: str,
-    ) -> tuple[EnvironmentHandle, Path | None, Path | None]:
-        target = plan.request.target
-        assert target.kind == "docker"
-        image = target.image or (
-            f"trtmc-devtoolkit:{plan.cohort.id}-{plan.architecture}-{plan.image_fingerprint[:12]}"
-        )
-        if target.image is None:
-            self._ensure_image(plan, image)
-        container = target.container_name or self._default_container_name(plan)
-        self._ensure_container(plan, image, container, sm)
-        observation = self._verify_container(plan, container, sm)
-        wheel = self._build_install(plan, container, sm)
-        handle = self._verify_install(plan, image, container, observation)
-        return handle, wheel, None
-
-    def _ensure_image(self, plan: PreparationPlan, image: str) -> None:
-        target = plan.request.target
-        assert target.kind == "docker"
-        expected = plan.image_fingerprint or ""
-        inspect = self.runner.run(
-            ["docker", "image", "inspect", image],
-            cwd=self.repository,
-            check=False,
-            capture_output=True,
-        )
-        if inspect.returncode == 0 and not target.rebuild_image:
-            actual = command_output(
-                self.runner,
-                [
-                    "docker",
-                    "image",
-                    "inspect",
-                    "--format",
-                    f'{{{{ index .Config.Labels "{IMAGE_FINGERPRINT_LABEL}" }}}}',
-                    image,
-                ],
-                cwd=self.repository,
-            )
-            if actual == expected:
-                return
-        contract = plan.cohort.architectures[plan.architecture]
-        self.runner.run(
-            [
-                "docker",
-                "build",
-                "--file",
-                self.repository / contract.dockerfile,
-                "--tag",
-                image,
-                "--label",
-                f"{IMAGE_FINGERPRINT_LABEL}={expected}",
-                self.repository / contract.docker_context,
-            ],
-            cwd=self.repository,
-        )
-
-    def _default_container_name(self, plan: PreparationPlan) -> str:
-        workspace = re.sub(r"[^a-z0-9]+", "-", self.repository.name.lower()).strip("-")
-        return f"trtmc-devtoolkit-{workspace}-{plan.environment_id[:8]}"
-
-    def _ensure_container(
-        self,
-        plan: PreparationPlan,
-        image: str,
-        container: str,
-        sm: str,
-    ) -> None:
-        if CONTAINER_NAME.fullmatch(container) is None:
-            raise DevToolkitError(f"Invalid Docker container name: {container!r}")
-        target = plan.request.target
-        assert target.kind == "docker"
-        inspect = self.runner.run(
-            ["docker", "container", "inspect", container],
-            cwd=self.repository,
-            check=False,
-            capture_output=True,
-        )
-        if inspect.returncode == 0:
-            owner = command_output(
-                self.runner,
-                [
-                    "docker",
-                    "container",
-                    "inspect",
-                    "--format",
-                    f'{{{{ index .Config.Labels "{ENVIRONMENT_LABEL}" }}}}',
-                    container,
-                ],
-                cwd=self.repository,
-            )
-            if owner != plan.environment_id:
-                raise DevToolkitError(
-                    f"Docker container {container} already exists and is not owned by "
-                    f"this preparation plan (label={owner or 'missing'})"
-                )
-            self.runner.run(["docker", "start", container], cwd=self.repository)
-            return
-
-        cache_root = plan.state_dir.parent / "cache"
-        hf_cache = Path(target.hf_cache).resolve() if target.hf_cache else cache_root / "hf"
-        storage_root = (
-            Path(target.storage_root).resolve() if target.storage_root else cache_root / "storage"
-        )
-        hf_cache.mkdir(parents=True, exist_ok=True)
-        storage_root.mkdir(parents=True, exist_ok=True)
-        command = [
-            "docker",
-            "run",
-            "--detach",
-            "--name",
-            container,
-            "--label",
-            f"{ENVIRONMENT_LABEL}={plan.environment_id}",
-            "--gpus",
-            f"device={target.gpu}",
-            "--ipc=host",
-            "--mount",
-            f"type=bind,source={self.repository},target=/workspace/tensorrt-model-connect",
-            "--mount",
-            f"type=bind,source={plan.state_dir},target=/trtmc-devtoolkit-run",
-            "--mount",
-            f"type=bind,source={hf_cache},target=/root/.cache/huggingface/hub",
-            "--mount",
-            f"type=bind,source={storage_root},target=/trtmc-storage",
-            "--workdir",
-            "/workspace/tensorrt-model-connect",
-            "--env",
-            f"TRTMC_SM={sm}",
-            "--env",
-            "ENGINE_DIR=/trtmc-storage/engines",
-        ]
-        for name in target.forward_environment:
-            if os.environ.get(name):
-                command.extend(["--env", name])
-        command.extend([image, "sleep", "infinity"])
-        self.runner.run(
-            command,
-            cwd=self.repository,
-        )
-
-    def _verify_container(
-        self, plan: PreparationPlan, container: str, sm: str
-    ) -> ToolchainObservation:
-        contract = plan.cohort.architectures[plan.architecture]
-        header = Path(contract.tensorrt_include_dir) / "NvInferVersion.h"
-        script = (
-            "import ctypes, json, sys, tensorrt; from pathlib import Path; "
-            f"assert tensorrt.__version__ == {plan.cohort.tensorrt_version!r}; "
-            f"lib=ctypes.CDLL({str(Path(contract.tensorrt_library_dir) / 'libnvinfer.so')!r}); "
-            "names=('Major','Minor','Patch','Build'); "
-            "funcs=[getattr(lib, f'getInferLib{name}Version') for name in names]; "
-            "[setattr(func, 'restype', ctypes.c_int32) for func in funcs]; "
-            f"text=Path({str(header)!r}).read_text(); "
-            "python=f'{sys.version_info.major}.{sys.version_info.minor}'; "
-            "print(json.dumps({'python': python, 'tensorrt_python': tensorrt.__version__, "
-            "'tensorrt_native': '.'.join(str(func()) for func in funcs), "
-            "'header_text': text}))"
-        )
-        output = command_output(
-            self.runner,
-            _docker_exec(container, ["python", "-c", script]),
-            cwd=self.repository,
-        )
-        try:
-            payload = json.loads(output)
-            python_version = payload["python"]
-            tensorrt_python = payload["tensorrt_python"]
-            tensorrt_native = payload["tensorrt_native"]
-            header_text = payload["header_text"]
-        except (json.JSONDecodeError, KeyError, TypeError) as error:
-            raise DevToolkitError(f"Container TensorRT probe returned invalid data: {error}") from error
-        if not all(
-            isinstance(value, str)
-            for value in (python_version, tensorrt_python, tensorrt_native, header_text)
-        ):
-            raise DevToolkitError("Container TensorRT probe returned non-text data")
-        tensorrt_headers = tensorrt_header_version_text(header_text, header)
-        expected = (
-            plan.request.python_version,
-            plan.cohort.tensorrt_version,
-            plan.cohort.tensorrt_version,
-            plan.cohort.tensorrt_version,
-        )
-        observed = (python_version, tensorrt_python, tensorrt_native, tensorrt_headers)
-        if observed != expected:
-            raise DevToolkitError(
-                "Container TensorRT Python/native/header mismatch: "
-                f"expected {expected}, got {observed}"
-            )
-        nvcc_output = command_output(
-            self.runner,
-            _docker_exec(container, ["nvcc", "--version"]),
-            cwd=self.repository,
-        )
-        match = CUDA_RELEASE.search(nvcc_output)
-        if match is None or match.group(1) != plan.cohort.cuda_version:
-            raise DevToolkitError(
-                f"Container CUDA mismatch: expected {plan.cohort.cuda_version}, got {nvcc_output}"
-            )
-        observed_sm = (
-            command_output(
-                self.runner,
-                _docker_exec(
-                    container,
-                    [
-                        "nvidia-smi",
-                        "--query-gpu=compute_cap",
-                        "--format=csv,noheader,nounits",
-                    ],
-                ),
-                cwd=self.repository,
-            )
-            .splitlines()[0]
-            .replace(".", "")
-            .strip()
-        )
-        if observed_sm != sm:
-            raise DevToolkitError(f"Container sees SM {observed_sm}; host selected SM {sm}")
-        return ToolchainObservation(
-            python_version=python_version,
-            cuda_version=match.group(1),
-            tensorrt_python_version=tensorrt_python,
-            tensorrt_native_version=tensorrt_native,
-            tensorrt_header_version=tensorrt_headers,
-            tensorrt_include_dir=contract.tensorrt_include_dir,
-            tensorrt_library=str(Path(contract.tensorrt_library_dir) / "libnvinfer.so"),
-        )
-
-    def _build_install(
-        self,
-        plan: PreparationPlan,
-        container: str,
-        sm: str,
-    ) -> Path | None:
-        state = _container_path(plan, plan.state_dir)
-        if plan.request.mode == "development":
-            build = state / f"build-sm{sm}"
-            self.runner.run(
-                _docker_exec(
-                    container,
-                    [
-                        "python",
-                        "-m",
-                        "pip",
-                        "install",
-                        "--no-deps",
-                        "-e",
-                        ".",
-                        "-C",
-                        "py-only=true",
-                    ],
-                ),
-                cwd=self.repository,
-            )
-            self.runner.run(
-                _docker_exec(
-                    container,
-                    [
-                        "cmake",
-                        "-S",
-                        ".",
-                        "-B",
-                        build,
-                        "-G",
-                        "Ninja",
-                        "-DCMAKE_BUILD_TYPE=Release",
-                        f"-DCMAKE_CUDA_ARCHITECTURES={sm}-real",
-                        "-DTRTMC_BUILD_BACKEND_TRT=ON",
-                        "-DTRTMC_BUILD_BACKEND_RTX=OFF",
-                        "-DTRTMC_BUILD_TESTS=OFF",
-                        "-DTRTMC_BUILD_BENCHMARKS=OFF",
-                        "-DTRTMC_ENABLE_LIBTORCH_MULTINOMIAL=OFF",
-                    ],
-                ),
-                cwd=self.repository,
-            )
-            self.runner.run(
-                _docker_exec(
-                    container,
-                    [
-                        "cmake",
-                        "--build",
-                        build,
-                        "--parallel",
-                        str(max(1, os.cpu_count() or 1)),
-                        "--target",
-                        "trtmc",
-                        "trtmc_backend_trt",
-                        "trtmc_model_plugins",
-                    ],
-                ),
-                cwd=self.repository,
-            )
-            return None
-
-        environment = {
-            "TRTMC_PACKAGE_TENSORRT_VERSION": plan.cohort.tensorrt_version,
-            "TRTMC_PACKAGE_PYTHON_TAGS": _python_tag(plan.request.python_version),
-            "TRTMC_PACKAGE_WHEEL_ARCH": plan.cohort.architectures[plan.architecture].wheel_platform,
-            "TRTMC_PACKAGE_BUILD_ROOT": str(state / "package-build"),
-            "TRTMC_CI_STATE_DIR": str(state / "ci-state"),
-        }
-        self.runner.run(
-            _docker_exec(
-                container,
-                ["python", "-m", "tools.ci", "pipeline", "package"],
-                environment=environment,
-            ),
-            cwd=self.repository,
-        )
-        return self._select_wheel(plan)
-
-    def _select_wheel(self, plan: PreparationPlan) -> Path:
-        tag = _python_tag(plan.request.python_version)
-        platform = plan.cohort.architectures[plan.architecture].wheel_platform
-        wheels = sorted((self.repository / "dist").glob(f"*-{tag}-none-{platform}.whl"))
-        if len(wheels) != 1:
-            raise DevToolkitError(f"Expected one built wheel for {tag}/{platform}, found {wheels}")
-        return wheels[0]
-
-    def _verify_install(
-        self,
-        plan: PreparationPlan,
-        image: str,
-        container: str,
-        observation: ToolchainObservation,
-    ) -> EnvironmentHandle:
-        state = _container_path(plan, plan.state_dir)
-        if plan.request.mode == "development":
-            build = state / f"build-sm{self._container_sm(container)}"
-            trtmc = str(build / "trtmc")
-            environment = {"TRTMC_MODEL_PLUGIN_DIR": str(build / "models")}
-        else:
-            trtmc = "/opt/venv/bin/trtmc"
-            environment = {}
-        self.runner.run(
-            _docker_exec(container, [trtmc, "version"], environment=environment),
-            cwd=self.repository,
-        )
-        self.runner.run(
-            _docker_exec(
-                container,
-                [
-                    "python",
-                    "-c",
-                    "import tensorrt_model_connect; print(tensorrt_model_connect.__file__)",
-                ],
-                environment=environment,
-            ),
-            cwd=self.repository,
-        )
-        return EnvironmentHandle(
-            kind="docker",
-            fingerprint=plan.environment_id,
-            trtmc=trtmc,
-            python="/opt/venv/bin/python",
-            activate_command=f"docker exec -it {container} bash",
-            image_ref=image,
-            container_name=container,
-            environment=environment,
-            observation=observation,
-        )
-
-    def _container_sm(self, container: str) -> str:
-        return command_output(
-            self.runner,
-            _docker_exec(container, ["sh", "-c", 'printf %s "$TRTMC_SM"']),
-            cwd=self.repository,
-        )
-
-
-class LocalEnvironment:
-    def __init__(self, repository: Path, runner: Runner):
-        self.repository = repository.resolve()
-        self.runner = runner
-
-    def prepare(
-        self,
-        plan: PreparationPlan,
-        *,
-        sm: str,
-    ) -> tuple[EnvironmentHandle, Path | None, Path | None]:
-        target = plan.request.target
-        assert target.kind == "local"
-        venv = plan.state_dir / "venv"
-        python = venv / "bin" / "python"
-        if not python.is_file():
-            venv.parent.mkdir(parents=True, exist_ok=True)
-            command: list[str | Path] = [target.python, "-m", "venv"]
-            if target.dependency_mode == "system":
-                command.append("--system-site-packages")
-            command.append(venv)
-            self.runner.run(command, cwd=self.repository)
-        provisioner = None
-        managed = None
-        if target.dependency_mode == "managed":
-            provisioner = ManagedLocalProvisioner(self.repository, self.runner)
-            managed = provisioner.prepare(plan, venv)
-            environment = managed.environment(venv, gpu=target.gpu)
-        else:
-            contract = plan.cohort.architectures[plan.architecture]
-            tensorrt_include_dir, tensorrt_library = system_tensorrt_paths(contract)
-            environment = {
-                "PATH": f"{venv / 'bin'}:{os.environ.get('PATH', '')}",
-                "VIRTUAL_ENV": str(venv),
-                "CUDA_VISIBLE_DEVICES": target.gpu,
-                "TRTMC_TRT_INCLUDE_DIR": str(tensorrt_include_dir),
-                "TRTMC_TRT_LIBRARY": str(tensorrt_library),
-                "TRTMC_TRT_LIBRARY_DIR": str(tensorrt_library.parent),
-            }
-        wheel = self._build_install(plan, python, environment, sm)
-        if provisioner is not None and managed is not None:
-            observation = provisioner.verify(plan, python, managed)
-        else:
-            observation = observe_local_toolchain(
-                self.runner,
-                repository=self.repository,
-                python=python,
-                nvcc="nvcc",
-                tensorrt_include_dir=tensorrt_include_dir,
-                tensorrt_library=tensorrt_library,
-                environment=environment,
-            )
-            require_toolchain_observation(
-                observation,
-                tensorrt=plan.cohort.tensorrt_version,
-                cuda=plan.cohort.cuda_version,
-                python=plan.request.python_version,
-            )
-        if plan.request.mode == "development":
-            build = plan.state_dir / f"build-sm{sm}"
-            trtmc = build / "trtmc"
-            runtime_env = _development_runtime_environment(environment, build)
-        else:
-            trtmc = venv / "bin" / "trtmc"
-            runtime_env = environment
-        activation = plan.state_dir / "activate.sh"
-        _write_local_activation(activation, venv, runtime_env)
-        self.runner.run([trtmc, "version"], cwd=self.repository, env=runtime_env)
-        self.runner.run(
-            [python, "-c", "import tensorrt_model_connect; print(tensorrt_model_connect.__file__)"],
-            cwd=self.repository,
-            env=runtime_env,
-        )
-        self.runner.run(
-            [python, "-m", "pip", "check"],
-            cwd=self.repository,
-            env=runtime_env,
-        )
-        handle = EnvironmentHandle(
-            kind="local",
-            fingerprint=plan.environment_id,
-            trtmc=str(trtmc),
-            python=str(python),
-            activate_command=f". {shlex.quote(str(activation))}",
-            environment=runtime_env,
-            observation=observation,
-        )
-        return handle, wheel, None
-
-    def _build_install(
-        self,
-        plan: PreparationPlan,
-        python: Path,
-        environment: dict[str, str],
-        sm: str,
-    ) -> Path | None:
-        contract = plan.cohort.architectures[plan.architecture]
-        if plan.request.mode == "development":
-            build = plan.state_dir / f"build-sm{sm}"
-            editable = [python, "-m", "pip", "install"]
-            if plan.request.target.dependency_mode == "system":
-                editable.append("--no-deps")
-            editable.extend(["-e", ".", "-C", "py-only=true"])
-            editable_environment = {
-                **environment,
-                "TRTMC_PACKAGE_TENSORRT_VERSION": plan.cohort.tensorrt_version,
-            }
-            self.runner.run(
-                editable,
-                cwd=self.repository,
-                env=editable_environment,
-            )
-            if plan.request.target.dependency_mode == "system":
-                requirements = _installed_base_requirements(
-                    python,
-                    self.runner,
-                    repository=self.repository,
-                    environment=editable_environment,
-                )
-                if requirements:
-                    self.runner.run(
-                        [python, "-m", "pip", "install", *requirements],
-                        cwd=self.repository,
-                        env=editable_environment,
-                    )
-            trt_include = environment.get("TRTMC_TRT_INCLUDE_DIR", contract.tensorrt_include_dir)
-            trt_library = environment.get(
-                "TRTMC_TRT_LIBRARY",
-                f"{contract.tensorrt_library_dir}/libnvinfer.so",
-            )
-            configure: list[str | Path] = [
-                "cmake",
-                "-S",
-                ".",
-                "-B",
-                build,
-                "-G",
-                "Ninja",
-                "-DCMAKE_BUILD_TYPE=Release",
-                f"-DCMAKE_CUDA_ARCHITECTURES={sm}-real",
-                f"-DTRTMC_TRT_INCLUDE_DIR={trt_include}",
-                f"-DTRTMC_TRT_LIBRARY={trt_library}",
-            ]
-            if environment.get("CUDA_HOME"):
-                configure.extend(
-                    [
-                        f"-DCMAKE_CUDA_COMPILER={environment['CUDA_HOME']}/bin/nvcc",
-                        f"-DTRTMC_CUDA_INCLUDE_DIR={environment['TRTMC_CUDA_INCLUDE_DIR']}",
-                        f"-DTRTMC_CUDART_LIBRARY={environment['TRTMC_CUDART_LIBRARY']}",
-                        f"-DTRTMC_CUBLAS_LIBRARY={environment['TRTMC_CUBLAS_LIBRARY']}",
-                    ]
-                )
-            configure.extend(
-                [
-                    "-DTRTMC_BUILD_BACKEND_TRT=ON",
-                    "-DTRTMC_BUILD_BACKEND_RTX=OFF",
-                    "-DTRTMC_BUILD_TESTS=OFF",
-                    "-DTRTMC_BUILD_BENCHMARKS=OFF",
-                    "-DTRTMC_ENABLE_LIBTORCH_MULTINOMIAL=OFF",
-                ]
-            )
-            self.runner.run(
-                configure,
-                cwd=self.repository,
-                env=environment,
-            )
-            self.runner.run(
-                [
-                    "cmake",
-                    "--build",
-                    build,
-                    "--parallel",
-                    str(max(1, os.cpu_count() or 1)),
-                    "--target",
-                    "trtmc",
-                    "trtmc_backend_trt",
-                    "trtmc_model_plugins",
-                ],
-                cwd=self.repository,
-                env=environment,
-            )
-            return None
-        package_env = {
-            **environment,
-            "TRTMC_PACKAGE_TENSORRT_VERSION": plan.cohort.tensorrt_version,
-            "TRTMC_PACKAGE_PYTHON_TAGS": _python_tag(plan.request.python_version),
-            "TRTMC_PACKAGE_WHEEL_ARCH": contract.wheel_platform,
-            "TRTMC_PACKAGE_BUILD_ROOT": str(plan.state_dir / "package-build"),
-            "TRTMC_CI_STATE_DIR": str(plan.state_dir / "ci-state"),
-        }
-        self.runner.run(
-            [python, "-m", "tools.ci", "pipeline", "package"],
-            cwd=self.repository,
-            env=package_env,
-        )
-        wheels = sorted(
-            (self.repository / "dist").glob(
-                f"*-{_python_tag(plan.request.python_version)}-none-{contract.wheel_platform}.whl"
-            )
-        )
-        if len(wheels) != 1:
-            raise DevToolkitError(f"Expected one installed wheel, found {wheels}")
-        return wheels[0]
