+++ scripts/devToolkit/trtmc_devtoolkit/toolchain.py
@@ -1,26 +1,14 @@
 # SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-"""User-owned CUDA/TensorRT toolchain provisioning for LocalTarget."""
+"""Toolchain observation helpers shared by built-in providers."""
 
 from __future__ import annotations
 
-import hashlib
-import os
 import re
-import urllib.error
-import urllib.parse
-import urllib.request
-from dataclasses import dataclass
 from pathlib import Path
 
-from .models import (
-    ArchitectureContract,
-    DevToolkitError,
-    DownloadArtifact,
-    PreparationPlan,
-    ToolchainObservation,
-)
+from .models import DevToolkitError, ToolchainObservation
 from .runner import Runner, command_output
 
 
@@ -43,21 +31,6 @@ def tensorrt_header_version_text(text: str, source: str | Path) -> str:
 
 def tensorrt_header_version(path: Path) -> str:
     return tensorrt_header_version_text(path.read_text(encoding="utf-8"), path)
-
-
-def system_tensorrt_paths(contract: ArchitectureContract) -> tuple[Path, Path]:
-    include_dir = Path(
-        os.environ.get("TRTMC_TRT_INCLUDE_DIR")
-        or os.environ.get("TRT_INC_DIR")
-        or contract.tensorrt_include_dir
-    )
-    library_dir = Path(
-        os.environ.get("TRTMC_TRT_LIBRARY_DIR")
-        or os.environ.get("TRT_LIB_DIR")
-        or contract.tensorrt_library_dir
-    )
-    library = Path(os.environ.get("TRTMC_TRT_LIBRARY") or library_dir / "libnvinfer.so")
-    return include_dir, library
 
 
 def observe_local_toolchain(
@@ -119,268 +92,3 @@ def observe_local_toolchain(
         tensorrt_library=str(tensorrt_library),
         cuda_root=str(cuda_root) if cuda_root is not None else None,
     )
-
-
-def require_toolchain_observation(
-    observation: ToolchainObservation,
-    *,
-    tensorrt: str,
-    cuda: str,
-    python: str,
-) -> None:
-    if observation.python_version != python:
-        raise DevToolkitError(f"Python must be {python}; got {observation.python_version}")
-    versions = {
-        observation.tensorrt_python_version,
-        observation.tensorrt_native_version,
-        observation.tensorrt_header_version,
-    }
-    if versions != {tensorrt}:
-        raise DevToolkitError(
-            f"TensorRT Python/native/header must all be {tensorrt}; got {sorted(versions)}"
-        )
-    if observation.cuda_version != cuda:
-        raise DevToolkitError(f"nvcc must report CUDA {cuda}; got {observation.cuda_version}")
-
-
-@dataclass(frozen=True)
-class ManagedToolchain:
-    cuda_root: Path
-    tensorrt_include_dir: Path
-    tensorrt_library_dir: Path
-    tensorrt_library: Path
-    cudart_library: Path
-    cublas_library: Path
-
-    def environment(self, venv: Path, *, gpu: str) -> dict[str, str]:
-        library_path = ":".join(
-            (
-                str(self.tensorrt_library_dir),
-                str(self.cuda_root / "lib"),
-            )
-        )
-        return {
-            "PATH": f"{venv / 'bin'}:{self.cuda_root / 'bin'}:{os.environ.get('PATH', '')}",
-            "VIRTUAL_ENV": str(venv),
-            "CUDA_VISIBLE_DEVICES": gpu,
-            "CUDA_HOME": str(self.cuda_root),
-            "CUDA_PATH": str(self.cuda_root),
-            "CUDAToolkit_ROOT": str(self.cuda_root),
-            "TRTMC_CUDA_INCLUDE_DIR": str(self.cuda_root / "include"),
-            "TRTMC_CUDART_LIBRARY": str(self.cudart_library),
-            "TRTMC_CUBLAS_LIBRARY": str(self.cublas_library),
-            "TRTMC_TRT_INCLUDE_DIR": str(self.tensorrt_include_dir),
-            "TRTMC_TRT_LIBRARY": str(self.tensorrt_library),
-            "TRTMC_TRT_LIBRARY_DIR": str(self.tensorrt_library_dir),
-            "TRT_INC_DIR": str(self.tensorrt_include_dir),
-            "TRT_LIB_DIR": str(self.tensorrt_library_dir),
-            "LD_LIBRARY_PATH": library_path,
-        }
-
-
-class ManagedLocalProvisioner:
-    def __init__(self, repository: Path, runner: Runner):
-        self.repository = repository.resolve()
-        self.runner = runner
-
-    def prepare(self, plan: PreparationPlan, venv: Path) -> ManagedToolchain:
-        python = venv / "bin" / "python"
-        requirements = [
-            f"{package.name}=={package.version}"
-            for package in plan.cohort.managed_local.python_packages
-        ]
-        self.runner.run(
-            [
-                python,
-                "-m",
-                "pip",
-                "install",
-                "--disable-pip-version-check",
-                *requirements,
-            ],
-            cwd=self.repository,
-        )
-        self.runner.run(
-            [python, "-m", "pip", "check"],
-            cwd=self.repository,
-        )
-
-        contract = plan.cohort.architectures[plan.architecture]
-        include_dir = self._prepare_headers(plan, contract.tensorrt_headers)
-        cuda_root = Path(
-            command_output(
-                self.runner,
-                [
-                    python,
-                    "-c",
-                    (
-                        "from pathlib import Path; import nvidia.cu13; "
-                        "print(Path(next(iter(nvidia.cu13.__path__))).resolve())"
-                    ),
-                ],
-                cwd=self.repository,
-            )
-        )
-        tensorrt_library_dir = Path(
-            command_output(
-                self.runner,
-                [
-                    python,
-                    "-c",
-                    (
-                        "import importlib.metadata as m; from pathlib import Path; "
-                        "print(Path(m.distribution('tensorrt_cu13_libs')"
-                        ".locate_file('tensorrt_libs')).resolve())"
-                    ),
-                ],
-                cwd=self.repository,
-            )
-        )
-        self._require_directory(cuda_root, "managed CUDA root")
-        self._require_directory(tensorrt_library_dir, "managed TensorRT library directory")
-
-        cudart = self._ensure_linker_name(cuda_root / "lib", "libcudart.so", "13")
-        cublas = self._ensure_linker_name(cuda_root / "lib", "libcublas.so", "13")
-        for linker_name, major in (
-            ("libcublasLt.so", "13"),
-            ("libcurand.so", "10"),
-        ):
-            self._ensure_linker_name(cuda_root / "lib", linker_name, major)
-        tensorrt = self._ensure_linker_name(
-            tensorrt_library_dir,
-            "libnvinfer.so",
-            plan.cohort.tensorrt_version.split(".", 1)[0],
-        )
-        for linker_name in (
-            "libnvinfer_plugin.so",
-            "libnvonnxparser.so",
-        ):
-            self._ensure_linker_name(
-                tensorrt_library_dir,
-                linker_name,
-                plan.cohort.tensorrt_version.split(".", 1)[0],
-            )
-
-        toolchain = ManagedToolchain(
-            cuda_root=cuda_root,
-            tensorrt_include_dir=include_dir,
-            tensorrt_library_dir=tensorrt_library_dir,
-            tensorrt_library=tensorrt,
-            cudart_library=cudart,
-            cublas_library=cublas,
-        )
-        return toolchain
-
-    def _prepare_headers(
-        self,
-        plan: PreparationPlan,
-        artifact: DownloadArtifact,
-    ) -> Path:
-        cache = plan.state_dir.parent / "downloads"
-        filename = Path(urllib.parse.urlparse(artifact.url).path).name
-        if not filename:
-            raise DevToolkitError(f"TensorRT header URL has no filename: {artifact.url}")
-        archive = cache / f"{artifact.sha256[:16]}-{filename}"
-        self._download(artifact, archive)
-        root = plan.state_dir / "managed-toolchain" / "tensorrt-headers"
-        marker = root / ".artifact-sha256"
-        if not marker.is_file() or marker.read_text(encoding="utf-8").strip() != artifact.sha256:
-            root.mkdir(parents=True, exist_ok=True)
-            self.runner.run(
-                ["dpkg-deb", "--extract", archive, root],
-                cwd=self.repository,
-            )
-            marker.write_text(artifact.sha256 + "\n", encoding="utf-8")
-        matches = sorted(root.glob("usr/include/*-linux-gnu/NvInferVersion.h"))
-        if len(matches) != 1:
-            raise DevToolkitError(
-                f"Expected one NvInferVersion.h in managed header artifact, found {matches}"
-            )
-        return matches[0].parent
-
-    def _download(self, artifact: DownloadArtifact, destination: Path) -> None:
-        if destination.is_file() and self._sha256(destination) == artifact.sha256:
-            return
-        destination.parent.mkdir(parents=True, exist_ok=True)
-        partial = destination.with_suffix(destination.suffix + f".partial-{os.getpid()}")
-        digest = hashlib.sha256()
-        try:
-            with urllib.request.urlopen(artifact.url, timeout=120) as response:
-                with partial.open("wb") as stream:
-                    while chunk := response.read(1024 * 1024):
-                        digest.update(chunk)
-                        stream.write(chunk)
-            actual = digest.hexdigest()
-            if actual != artifact.sha256:
-                raise DevToolkitError(
-                    f"TensorRT header artifact checksum mismatch: expected "
-                    f"{artifact.sha256}, got {actual}"
-                )
-            partial.replace(destination)
-        except (OSError, urllib.error.URLError) as error:
-            raise DevToolkitError(
-                f"Could not download TensorRT header artifact {artifact.url}: {error}"
-            ) from error
-        finally:
-            if partial.exists():
-                partial.unlink()
-
-    @staticmethod
-    def _sha256(path: Path) -> str:
-        digest = hashlib.sha256()
-        with path.open("rb") as stream:
-            while chunk := stream.read(1024 * 1024):
-                digest.update(chunk)
-        return digest.hexdigest()
-
-    @staticmethod
-    def _require_directory(path: Path, name: str) -> None:
-        if not path.is_dir():
-            raise DevToolkitError(f"Could not locate {name}: {path}")
-
-    @staticmethod
-    def _ensure_linker_name(directory: Path, linker_name: str, major: str) -> Path:
-        linker = directory / linker_name
-        versioned = directory / f"{linker_name}.{major}"
-        if not versioned.is_file():
-            raise DevToolkitError(f"Managed toolchain is missing {versioned}")
-        if linker.is_symlink() and linker.resolve() == versioned.resolve():
-            return linker
-        if linker.exists() or linker.is_symlink():
-            raise DevToolkitError(
-                f"Managed toolchain linker name does not select {versioned}: {linker}"
-            )
-        linker.symlink_to(versioned.name)
-        return linker
-
-    def verify(
-        self,
-        plan: PreparationPlan,
-        python: Path,
-        toolchain: ManagedToolchain,
-    ) -> ToolchainObservation:
-        environment = toolchain.environment(
-            python.parent.parent,
-            gpu=plan.request.target.gpu,
-        )
-        observation = observe_local_toolchain(
-            self.runner,
-            repository=self.repository,
-            python=python,
-            nvcc=toolchain.cuda_root / "bin" / "nvcc",
-            tensorrt_include_dir=toolchain.tensorrt_include_dir,
-            tensorrt_library=toolchain.tensorrt_library,
-            environment=environment,
-            cuda_root=toolchain.cuda_root,
-        )
-        require_toolchain_observation(
-            observation,
-            tensorrt=plan.cohort.tensorrt_version,
-            cuda=plan.cohort.cuda_version,
-            python=plan.request.python_version,
-        )
-        return observation
-
-    @staticmethod
-    def _header_version(path: Path) -> str:
-        return tensorrt_header_version(path)
